add defalt.jsの処理をファイル毎に分割
default.jsは削除済。
